Remove JobSerializer
