+++ src/Model/JobStatus.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace App\Model;
 
+use App\Entity\Job;
 use App\Enum\CompilationState;
 use App\Enum\EventDeliveryState;
 use App\Enum\ExecutionState;
@@ -11,18 +12,19 @@ use App\Enum\ExecutionState;
 class JobStatus implements \JsonSerializable
 {
     /**
-     * @param array<mixed> $serializedJob
-     * @param string[]     $sourcePaths
-     * @param array<mixed> $serializedTests
+     * @param string[]                                            $sourcePaths
+     * @param array<mixed>                                        $serializedTests
+     * @param array<int, array{label: string, reference: string}> $testReferences
      */
     public function __construct(
-        private readonly array $serializedJob,
+        private readonly Job $job,
         private readonly string $reference,
         private readonly array $sourcePaths,
         private readonly CompilationState $compilationState,
         private readonly ExecutionState $executionState,
         private readonly EventDeliveryState $eventDeliveryState,
-        private readonly array $serializedTests
+        private readonly array $serializedTests,
+        private readonly array $testReferences,
     ) {
     }
 
@@ -31,16 +33,18 @@ class JobStatus implements \JsonSerializable
      */
     public function jsonSerialize(): array
     {
-        return array_merge(
-            $this->serializedJob,
-            [
-                'reference' => $this->reference,
-                'sources' => $this->sourcePaths,
-                'compilation_state' => $this->compilationState->value,
-                'execution_state' => $this->executionState->value,
-                'event_delivery_state' => $this->eventDeliveryState->value,
-                'tests' => $this->serializedTests,
-            ]
-        );
+        return [
+            'label' => $this->job->getLabel(),
+            'event_delivery_url' => $this->job->getEventDeliveryUrl(),
+            'maximum_duration_in_seconds' => $this->job->getMaximumDurationInSeconds(),
+            'test_paths' => $this->job->getTestPaths(),
+            'reference' => $this->reference,
+            'sources' => $this->sourcePaths,
+            'compilation_state' => $this->compilationState->value,
+            'execution_state' => $this->executionState->value,
+            'event_delivery_state' => $this->eventDeliveryState->value,
+            'tests' => $this->serializedTests,
+            'references' => $this->testReferences,
+        ];
     }
 }
